+++ src/causalitytests/tests_entropy_based/TransferEntropyCausalityTest.jl
@@ -4,6 +4,13 @@
 
 The supertype of all abstract and composite types representing a transfer 
 entropy causality test.
+
+Concrete subtypes are 
+
+- [`TransferOperatorGridTest`](@ref)
+- [`VisitationFrequencyTest`](@ref)
+- [`ApproximateSimplexIntersectionTest`](@ref)
+- [`ExactSimplexIntersectionTest`](@ref)
 """
 abstract type TransferEntropyCausalityTest <: EntropyBasedCausalityTest end