+++ docs/src/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog for CausalityTools.jl
 
+## Release v0.6.0
+
+### New features
+
+- Added `SMeasureTest` causality test.
+- Added tutorial for `SMeasureTest`.
+
+### Documentation
+
+- Improved documentation.
+
 ## Release v0.5.2
 
 ### Documentation